统一类型为int64

+++ app/repository/comment.go
@@ -4,8 +4,8 @@ import (
 	"tiktok/app/config"
 )
 
-func GetCommentCountByVideoId(videoId int) int {
+func GetCommentCountByVideoId(videoId int64) int64 {
 	var count int64
 	config.Db.Table("comments").Where("video_id = ?", videoId).Count(&count)
-	return int(count)
+	return int64(count)
 }
