+++ app/schema/comment.go
@@ -3,9 +3,9 @@ package schema
 import "time"
 
 type Comment struct {
-	Id          int    `json:"id"`
-	UserId      int    `json:"user_id"`
-	VideoId     int    `json:"video_id"`
+	Id          int64  `json:"id"`
+	UserId      int64  `json:"user_id"`
+	VideoId     int64  `json:"video_id"`
 	CommentText string `json:"comment_text"`
 	CreateDate  time.Time
 	ActionType  string `json:"action_type"`
