+++ app/vo/common.go
@@ -1,49 +1,49 @@
 package vo
 
 type Response struct {
-	StatusCode int32  `json:"status_code"`
+	StatusCode int64  `json:"status_code"`
 	StatusMsg  string `json:"status_msg,omitempty"`
 }
 
 type Video struct {
-	Id            int    `json:"id,omitempty"`
+	Id            int64  `json:"id,omitempty"`
 	Author        User   `json:"author"`
 	PlayUrl       string `json:"play_url" json:"play_url,omitempty"`
 	CoverUrl      string `json:"cover_url,omitempty"`
-	FavoriteCount int    `json:"favorite_count,omitempty"`
-	CommentCount  int    `json:"comment_count,omitempty"`
+	FavoriteCount int64  `json:"favorite_count,omitempty"`
+	CommentCount  int64  `json:"comment_count,omitempty"`
 	IsFavorite    bool   `json:"is_favorite,omitempty"`
 	Title         string `json:"title,omitempty"`
 }
 
 type Comment struct {
-	Id         int    `json:"id,omitempty"`
+	Id         int64  `json:"id,omitempty"`
 	User       User   `json:"user"`
 	Content    string `json:"content,omitempty"`
 	CreateDate string `json:"create_date,omitempty"`
 }
 
 type User struct {
-	Id            int    `json:"id"`
+	Id            int64  `json:"id"`
 	Name          string `json:"name"`
-	FollowCount   int    `json:"follow_count"`
-	FollowerCount int    `json:"follower_count"`
+	FollowCount   int64  `json:"follow_count"`
+	FollowerCount int64  `json:"follower_count"`
 	IsFollow      bool   `json:"is_follow"`
 }
 
 type Message struct {
-	Id         int    `json:"id,omitempty"`
+	Id         int64  `json:"id,omitempty"`
 	Content    string `json:"content,omitempty"`
 	CreateTime string `json:"create_time,omitempty"`
 }
 
 type MessageSendEvent struct {
-	UserId     int    `json:"user_id,omitempty"`
-	ToUserId   int    `json:"to_user_id,omitempty"`
+	UserId     int64  `json:"user_id,omitempty"`
+	ToUserId   int64  `json:"to_user_id,omitempty"`
 	MsgContent string `json:"msg_content,omitempty"`
 }
 
 type MessagePushEvent struct {
-	FromUserId int    `json:"user_id,omitempty"`
+	FromUserId int64  `json:"user_id,omitempty"`
 	MsgContent string `json:"msg_content,omitempty"`
 }
